requirements.txt
numpy~=1.20.3
numba~=0.54.1
python-mnist~=0.7